+++ scripts/libs/jansson.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/bash
+# 
+
+echo "---> Install jansson library";
+	curl -fsSL "http://amiga-projects.net/jansson_library_2.12.1_sdk.lha" -o /tmp/jansson.lha && \
+		lha -xfq2 jansson.lha && \
+		cp -r ./local/* ${OS4_SDK_PATH}/local/ && \
+		rm -rf /tmp/*;